Refactor file names to fit the convention
