+++ src/components/product-display.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Button } from "./button";
 import { Posts } from "./product-list-components";
-import styles from "../shopApp.module.css";
+import styles from "../shop-app.module.css";
 import { IProduct } from "../useShopApp";
 
 const countFavorite = (products: IProduct[]) => {
